Split utils, extended spectra and hdf5
